+++ src/algorithms/Neal8.hpp
@@ -30,7 +30,7 @@ public:
         const Eigen::MatrixXd &data_, const unsigned int init = 0) :
         Neal2<Hierarchy, Hypers, Mixture>::Neal2(hypers_, mixture_, data_,
         init) {
-        for(unsigned int i = 0; i < n_aux; i++){
+        for(size_t i = 0; i < n_aux; i++){
             aux_unique_values.push_back(this->unique_values[0]);
         }
     }
