+++ src/hierarchies/HierarchyNNIG.imp.hpp
@@ -13,7 +13,7 @@ void HierarchyNNIG<Hypers>::check_state_validity(){
 template<class Hypers> 
 Eigen::VectorXd HierarchyNNIG<Hypers>::like(const Eigen::MatrixXd &data){
     Eigen::VectorXd result(data.rows());
-    for(int i = 0; i < data.rows(); i++){
+    for(size_t i = 0; i < data.rows(); i++){
         result(i) = exp(stan::math::normal_lpdf(data(i,0), this->state[0](0,0),
             this->state[1](0,0)));
     }
@@ -41,7 +41,7 @@ Eigen::VectorXd HierarchyNNIG<Hypers>::eval_marg(const Eigen::MatrixXd &data){
         )+1) / (this->hypers->get_alpha0()*this->hypers->get_lambda()) );
    
     Eigen::VectorXd result(data.rows());
-    for(int i = 0; i < data.rows(); i++){
+    for(size_t i = 0; i < data.rows(); i++){
         result(i) = exp( stan::math::student_t_lpdf(data(i,0),
             2*this->hypers->get_alpha0(), this->hypers->get_mu0(), sigtilde) );
     }
